+++ src/app/personside/visittkort/header/Etiketter/Etiketter.tsx
@@ -2,12 +2,11 @@ import * as React from 'react';
 import styled from 'styled-components';
 
 import { Person } from '../../../../../models/person/person';
-import EtikettBase from 'nav-frontend-etiketter';
-import { Diskresjonskoder } from '../../../../../konstanter';
-import { Kodeverk } from '../../../../../models/kodeverk';
 import SikkerhetstiltakEtikett from './SikkerhetstiltakEtikett';
 import EgenAnsattEtikett from './EgenansattEtikettContainer';
 import VergemålEtikettContainer from './VergemålEtikettContainer';
+import DiskresjonskodeEtikett from './DiskresjonskodeEtikett';
+import TilrettelagtKommunikasjonsEtiketter from './TilrettelagtKommunikasjonsEtiketter';
 
 interface Props {
     person: Person;
@@ -20,25 +19,6 @@ const StyledEtikketter = styled.section`
     }
 `;
 
-function DiskresjonskodeEtikett(props: { diskresjonskode?: Kodeverk }) {
-    if (!props.diskresjonskode) {
-        return null;
-    }
-
-    switch (props.diskresjonskode.kodeRef) {
-        case Diskresjonskoder.STRENGT_FORTROLIG_ADRESSE:
-            return <EtikettBase type={'advarsel'}>Kode 6</EtikettBase>;
-        case Diskresjonskoder.FORTROLIG_ADRESSE:
-            return <EtikettBase type={'advarsel'}>Kode 7</EtikettBase>;
-        default:
-            return null;
-    }
-}
-
-function TilrettelagtKommunikasjonEtikett(props: { tilrettelagtKommunikasjon: Kodeverk }) {
-    return <EtikettBase type={'fokus'}>{props.tilrettelagtKommunikasjon.beskrivelse}</EtikettBase>;
-}
-
 function Etiketter({ person }: Props) {
     return (
         <StyledEtikketter role="region" aria-label="etiketter">
@@ -46,12 +26,9 @@ function Etiketter({ person }: Props) {
             <EgenAnsattEtikett />
             <SikkerhetstiltakEtikett sikkerhetstiltak={person.sikkerhetstiltak} />
             <VergemålEtikettContainer />
-            {person.tilrettelagtKomunikasjonsListe.map(tilrettelagtKommunikasjon => (
-                <TilrettelagtKommunikasjonEtikett
-                    key={tilrettelagtKommunikasjon.kodeRef}
-                    tilrettelagtKommunikasjon={tilrettelagtKommunikasjon}
-                />
-            ))}
+            <TilrettelagtKommunikasjonsEtiketter
+                tilrettelagtKomunikasjonsListe={person.tilrettelagtKomunikasjonsListe}
+            />
         </StyledEtikketter>
     );
 }
